Reorganize folders and namespaces
