+++ CircuitNotIncluded/Core/Runtime/InputRuntime.cs
@@ -1,20 +1,16 @@
-using KSerialization;
-using Newtonsoft.Json.Linq;
+using CircuitNotIncluded.Grammar;
 
-namespace CircuitNotIncluded.Structs;
+namespace CircuitNotIncluded.Core.Runtime;
 
 public class InputRuntime(SymbolTable symbolTable, string id, int cell) 
-	: CircuitPort(cell), ILogicEventReceiver
+	: PortRuntime(cell), ILogicEventReceiver
 {
 	public void OnLogicNetworkConnectionChanged(bool connected){ }
 	
 	public void ReceiveLogicEvent(int value){
 		symbolTable.SetValue(id, value);
 	}
-	
-	public int GetLogicCell(){
-		return cell;
-	}
-	
+
+	public int GetLogicCell() => base.GetLogicUICell();
 	public override LogicPortSpriteType GetLogicPortSpriteType() => LogicPortSpriteType.Input;
 }
