+++ CircuitNotIncluded/Core/Runtime/OutputRuntime.cs
@@ -1,24 +1,24 @@
 using CircuitNotIncluded.Grammar;
 using KSerialization;
-using Newtonsoft.Json.Linq;
 
-namespace CircuitNotIncluded.Structs;
+namespace CircuitNotIncluded.Core.Runtime;
 using EvaluateFunc = Func<SymbolTable, int>;
 
 [SerializationConfig(MemberSerialization.OptIn)]
 public class OutputRuntime(SymbolTable symbolTable, string expression, int cell) 
-	: CircuitPort(cell), ILogicEventSender
+	: PortRuntime(cell), ILogicEventSender
 {
 	private readonly EvaluateFunc evaluate = Compiler.Compile(expression);
 	private int logicValue;
 	
+	public void OnLogicNetworkConnectionChanged(bool connected){ }
+	
 	public void LogicTick(){
 		logicValue = evaluate(symbolTable);
 	}
-	
-	public void OnLogicNetworkConnectionChanged(bool connected){ }
-	public int GetLogicCell() => cell;
+
 	public int GetLogicValue() => logicValue;
 	
+	public int GetLogicCell() => base.GetLogicUICell();	
 	public override LogicPortSpriteType GetLogicPortSpriteType() => LogicPortSpriteType.Output;
 }
