+++ CircuitNotIncluded/Core/Structs/Validators/ExpHasValidIds.cs
@@ -2,7 +2,7 @@ using CircuitNotIncluded.Grammar;
 using CircuitNotIncluded.UI.Cells;
 using static CircuitNotIncluded.Grammar.ExpressionParser;
 
-namespace CircuitNotIncluded.UI.Validators;
+namespace CircuitNotIncluded.Core.Structs.Validators;
 
 public class ExpHasValidIdsValidator : BasePortValidator<OutputCellState> {
 	private readonly HashSet<string> invalidIds = [];
