+++ CircuitNotIncluded/Core/Structs/Validators/IdNotReservedWord.cs
@@ -1,7 +1,7 @@
 using CircuitNotIncluded.UI.Cells;
 using static CircuitNotIncluded.Grammar.ExpressionUtils;
 
-namespace CircuitNotIncluded.UI.Validators;
+namespace CircuitNotIncluded.Core.Structs.Validators;
 
 public class IdNotReservedWord : BasePortValidator<PortCellState> {
 	protected override bool DispatchErrorWhen(PortCellState port, ValidationContext ctx)
