+++ CircuitNotIncluded/UI/Cells/EmptyCellState.cs
@@ -1,4 +1,4 @@
-using CircuitNotIncluded.Structs.Ports;
+using CircuitNotIncluded.Core.Structs;
 using CircuitNotIncluded.Utils;
 using PeterHan.PLib.UI;
 using UnityEngine;
