+++ CircuitNotIncluded/UI/CircuitScreenManager.UI.cs
@@ -1,5 +1,5 @@
-using CircuitNotIncluded.Structs;
-using CircuitNotIncluded.Structs.Ports;
+using CircuitNotIncluded.Core;
+using CircuitNotIncluded.Core.Structs;
 using CircuitNotIncluded.UI.Cells;
 using CircuitNotIncluded.Utils;
 using PeterHan.PLib.Core;
